fix(SecureRng): Don't reset buffer on block state overflow
