+++ src/source/chacha/constants.rs
@@ -1,0 +1,1 @@
+pub(super) const INITIAL_STATE: &[u8; 16] = b"expand 32-byte k";
